feat : GH-52 Refactor auth module
