+++ core/network/src/main/kotlin/me/jerryokafor/ihenkiri/core/network/model/response/CreateGuestTokenResponse.kt
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2023 IheNkiri Project
+ * Copyright (c) 2024 IheNkiri Project
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,5 +30,6 @@ data class CreateGuestTokenResponse(
     @SerializedName("guest_session_id")
     val guestSessionId: String,
     val success: Boolean,
+    @SerializedName("expires_at")
     val expiresAt: String,
 )
