+++ core/network/src/main/kotlin/me/jerryokafor/ihenkiri/core/network/service/TVShowsListApi.kt
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2023 IheNkiri Project
+ * Copyright (c) 2024 IheNkiri Project
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,7 +30,7 @@ import me.jerryokafor.ihenkiri.core.network.model.response.PagedNetworkResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface TVSeriesListsApi {
+interface TVShowsListApi {
     /**
      * Get a list of TV shows airing today.
      *
