+++ core/network/src/main/resources/api-response/create-guest-token-200.json
@@ -1,0 +1,5 @@
+{
+  "guest_session_id" : "e25304417c5c67f66b7838517257a3e9",
+  "success" : true,
+  "expires_at" : "2023-11-22 21:06:30 UT"
+}
