+++ core/network/src/test/java/me/jerryokafor/core/network/service/AuthApiTest.kt
@@ -28,53 +28,85 @@ import kotlinx.coroutines.test.runTest
 import me.jerryokafor.core.network.util.enqueueResponse
 import me.jerryokafor.ihenkiri.core.network.Config
 import me.jerryokafor.ihenkiri.core.network.service.AuthApi
-import me.jerryokafor.ihenkiri.core.test.util.MockWebServerUtil.createMockedService
+import me.jerryokafor.ihenkiri.core.test.util.MockWebServerUtil
 import me.jerryokafor.ihenkiri.core.test.util.createAccessTokenRequest
 import me.jerryokafor.ihenkiri.core.test.util.createAccessTokenSuccessResponse
+import me.jerryokafor.ihenkiri.core.test.util.createGuestTokenResponse
 import me.jerryokafor.ihenkiri.core.test.util.createRequestTokenRequest
 import me.jerryokafor.ihenkiri.core.test.util.createRequestTokenSuccessResponse
+import okhttp3.mockwebserver.MockResponse
 import org.junit.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
 class AuthApiTest : BaseServiceTest() {
-    private val authApi =
-        createMockedService(mockWebServer, AuthApi::class.java)
+    private val authApi = MockWebServerUtil
+        .createMockedService(mockWebServer, AuthApi::class.java)
 
     @Test
-    fun `test createRequestToken(), returns request token if response is 200`() {
+    fun `test createRequestToken(), returns request token if response is 200`() = runTest {
         mockWebServer.enqueueResponse("create-request-token-200.json", 200)
 
-        runTest {
-            val requestBody = createRequestTokenRequest()
-            val response = authApi.createRequestToken(requestBody)
-            val expected = createRequestTokenSuccessResponse()
+        val requestBody = createRequestTokenRequest()
+        val response = authApi.createRequestToken(requestBody)
+        val expected = createRequestTokenSuccessResponse()
 
-            assertNotNull(response)
-            assertEquals(expected, response)
+        assertNotNull(response)
+        assertEquals(expected, response)
 
-            val recordedRequest = mockWebServer.takeRequest()
-            assertEquals(mockWebServer.requestCount, 1)
-            assertEquals("POST", recordedRequest.method)
-            assertEquals("/${Config.TMDB_API_V4}/auth/request_token", recordedRequest.path)
-        }
+        val recordedRequest = mockWebServer.takeRequest()
+        assertEquals(mockWebServer.requestCount, 1)
+        assertEquals("POST", recordedRequest.method)
+        assertEquals("/${Config.TMDB_API_V4}/auth/request_token", recordedRequest.path)
     }
 
     @Test
-    fun `test createAccessToken(), return access token if response is 200`() {
+    fun `test createAccessToken(), return access token if response is 200`() = runTest {
         mockWebServer.enqueueResponse("create-access-token-200.json", 200)
 
-        runTest {
-            val response = authApi.createAccessToken(createAccessTokenRequest())
-            val expected = createAccessTokenSuccessResponse()
+        val response = authApi.createAccessToken(createAccessTokenRequest())
+        val expected = createAccessTokenSuccessResponse()
 
-            assertNotNull(response)
-            assertEquals(expected, response)
+        assertNotNull(response)
+        assertEquals(expected, response)
 
-            val recordedRequest = mockWebServer.takeRequest()
-            assertEquals(mockWebServer.requestCount, 1)
-            assertEquals("POST", recordedRequest.method)
-            assertEquals("/${Config.TMDB_API_V4}/auth/access_token", recordedRequest.path)
-        }
+        val recordedRequest = mockWebServer.takeRequest()
+        assertEquals(mockWebServer.requestCount, 1)
+        assertEquals("POST", recordedRequest.method)
+        assertEquals("/${Config.TMDB_API_V4}/auth/access_token", recordedRequest.path)
+    }
+
+    @Test
+    fun `test createGuestSession(), return access token if response is 200`() = runTest {
+        mockWebServer.enqueueResponse("create-guest-token-200.json", 200)
+
+        val response = authApi.createGuestSession()
+        val expected = createGuestTokenResponse()
+
+        assertNotNull(response)
+        assertEquals(expected, response)
+
+        val recordedRequest = mockWebServer.takeRequest()
+        assertEquals(mockWebServer.requestCount, 1)
+        assertEquals("GET", recordedRequest.method)
+        assertEquals(
+            "/${Config.TMDB_API_V3}/authentication/guest_session/new",
+            recordedRequest.path,
+        )
+    }
+
+    @Test
+    fun `test deleteSession(), return success token if response is 200`() = runTest {
+        mockWebServer.enqueue(MockResponse().setResponseCode(200))
+
+        authApi.deleteSession()
+
+        val recordedRequest = mockWebServer.takeRequest()
+        assertEquals(mockWebServer.requestCount, 1)
+        assertEquals("DELETE", recordedRequest.method)
+        assertEquals(
+            "/${Config.TMDB_API_V4}/auth/access_token",
+            recordedRequest.path,
+        )
     }
 }
